requirements.txt
autopep8==1.5.5
Django==3.1.6
django-ckeditor==6.0.0
django-crispy-forms==1.11.0
django-js-asset==1.2.2
Pillow==8.1.0
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2